requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
mysqlclient==2.2.4
peewee==3.17.6
python-dotenv==1.0.1